feat(gha): publish developer package
Introduces a GitHub Actions workflow `publish-dev-package` which
publishes a developer package to the GitHub npm registry when commits
are pushed to the `main` branch. A developer package bears the target
release version but followed by the short commit hash of the commit used
to build the package. The actual operations are done in
`publish-package.yml`.

The workflow is also triggered by the `issue/2-gha-publish-dev` branch
for experiments. But this trigger must be removed when experiments are
completed.